# performance: fix performance for cases with sparsely connected entities