--- a/perf/.eslintrc
+++ b/perf/.eslintrc
@@ -2,5 +2,7 @@
   rules: {
     no-console: off,
     no-process-exit: off,
+    no-unused-vars: off,
+    no-empty: off
   }
 }