--- a/perf/.eslintrc
+++ b/perf/.eslintrc
@@ -1,8 +1,6 @@
 {
   rules: {
     no-console: off,
-    no-process-exit: off,
-    no-unused-vars: off,
-    no-empty: off
+    no-process-exit: off
   }
 }