--- a/perf/.eslintrc
+++ b/perf/.eslintrc
@@ -1,6 +1,6 @@
 {
   rules: {
     no-console: off,
-    no-process-exit: off
+    no-process-exit: off,
   }
 }