--- a/perf/N3Store-perf.js
+++ b/perf/N3Store-perf.js
@@ -104,12 +104,14 @@ console.timeEnd(TEST);
 TEST = `* Retrieving all ${dimQuads} quads`;
 console.time(TEST);
 for (const quad of store.match(undefined, undefined, undefined)) {
+  assert(quad);
 }
 console.timeEnd(TEST);
 
 TEST = '* Retrieving single by subject';
 console.time(TEST);
 for (const quad of store.match(prefix + 1, undefined, undefined)) {
+  assert(quad);
 }
 console.timeEnd(TEST);
 
@@ -117,12 +119,14 @@ console.timeEnd(TEST);
 TEST = '* Retrieving single by predicate';
 console.time(TEST);
 for (const quad of store.match(undefined, prefix + 1, undefined)) {
+  assert(quad);
 }
 console.timeEnd(TEST);
 
 TEST = '* Retrieving single by object';
 console.time(TEST);
 for (const quad of store.match(undefined, undefined, prefix + 1)) {
+  assert(quad);
 }
 console.timeEnd(TEST);
 
@@ -130,6 +134,7 @@ console.timeEnd(TEST);
 TEST = '* Retrieving single by subject-predicate';
 console.time(TEST);
 for (const quad of store.match(prefix + 1, prefix + 1, undefined)) {
+  assert(quad);
 }
 console.timeEnd(TEST);
 
@@ -137,12 +142,14 @@ console.timeEnd(TEST);
 TEST = '* Retrieving single by subject-object';
 console.time(TEST);
 for (const quad of store.match(prefix + 1, undefined, prefix + 1)) {
+  assert(quad);
 }
 console.timeEnd(TEST);
 
 TEST = '* Retrieving single by predicate-object';
 console.time(TEST);
 for (const quad of store.match(undefined, prefix + 1, prefix + 1)) {
+  assert(quad);
 }
 console.timeEnd(TEST);
 
@@ -150,5 +157,6 @@ console.timeEnd(TEST);
 TEST = '* Retrieving single by subject-predicate-object';
 console.time(TEST);
 for (const quad of store.match(prefix + 1, prefix + 1, prefix + 1)) {
+  assert(quad);
 }
 console.timeEnd(TEST);