--- a/src/N3Store.js
+++ b/src/N3Store.js
@@ -88,8 +88,7 @@ export default class N3Store {
   *_findInIndex(index0, key0, key1, key2, name0, name1, name2, graphId) {
     let tmp, index1, index2;
     // Depending on the number of variables, keys or reverse index are faster
-    const varCount = !key0 + !key1 + !key2,
-        entityKeys = varCount > 1 ? Object.keys(this._ids) : this._entities;
+    const entityKeys = !(key0 || key1 || key2) ? Object.keys(this._ids) : this._entities;
     const graph = termFromId(graphId, this._factory);
 
     // If a key is specified, use only that part of index 0.