Resolves #277

This PR improves the performance `match`ing against a pattern with one defined term on graphs that are more sparsely connected. Achieving a performance improvement of `60_000`x on the degenerate case of an unconnected graph.

I have added new performance tests to demonstrate this; using a dimension of 128 the results are

Before Changes
```
N3Store performance test
- Adding 2097152 triples to the default graph: 1.269s
* Memory usage for triples: 147MB
- Finding all 2097152 triples in the default graph 16384 times (0 variables): 5.652s
- Finding all 2097152 triples in the default graph 32768 times (1 variable): 1.646s
- Finding all 2097152 triples in the default graph 49152 times (2 variables): 1.854s

- Adding 1048576 quads: 870.612ms
* Memory usage for quads: 127MB
- Finding all 1048576 quads 131072 times: 1.389s

- Adding 1048576 with all different IRIs: 5.873s
* Retrieving all 1048576 quads: 1.460s
* Retrieving single by subject: 310.78ms
* Retrieving single by predicate: 313.62ms
* Retrieving single by object: 305.24ms
* Retrieving single by subject-predicate: 0.061ms
* Retrieving single by subject-object: 0.038ms
* Retrieving single by predicate-object: 0.04ms
* Retrieving single by subject-predicate-object: 0.209ms
```

After Changes
```
N3Store performance test
- Adding 2097152 triples to the default graph: 1.336s
* Memory usage for triples: 149MB
- Finding all 2097152 triples in the default graph 16384 times (0 variables): 5.319s
- Finding all 2097152 triples in the default graph 32768 times (1 variable): 1.504s
- Finding all 2097152 triples in the default graph 49152 times (2 variables): 1.817s

- Adding 1048576 quads: 829.416ms
* Memory usage for quads: 124MB
- Finding all 1048576 quads 131072 times: 1.357s

- Adding 1048576 with all different IRIs: 5.941s
* Retrieving all 1048576 quads: 1.554s
* Retrieving single by subject: 0.057ms
* Retrieving single by predicate: 0.112ms
* Retrieving single by object: 0.062ms
* Retrieving single by subject-predicate: 0.062ms
* Retrieving single by subject-object: 0.117ms
* Retrieving single by predicate-object: 0.084ms
* Retrieving single by subject-predicate-object: 0.107ms
```