@jeswr 2022-05-12
> I take it we are ready for a patch release now?

Yes @RubenVerborgh